Refactored Patch guts to script.
