+++ kCustomRadial/kCustomRadial.csproj
@@ -59,6 +59,7 @@
     <ItemGroup>
         <Compile Include="Mod\Harmony\Hello_Init.cs" />
         <Compile Include="Mod\Harmony\KProCustomRadial.cs" />
+        <Compile Include="Mod\Scripts\KRadial.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
     </ItemGroup>
     <ItemGroup>
